hoyt: dtype narrowing — stored waves int32/int8, leaf_pts int16
Working arrays stay int64 (one wave at a time); arithmetic consumers cast
on read (int8 tile in 9*tile would silently overflow under NEP 50).
PS/GID/uedge deliberately stay int64: they feed bincount/fancy-indexing in
the hot iterate loop, where numpy converts non-intp index arrays per call.

555006 cap-256 bit-identical; RSS 6.98->6.76 GiB, wall neutral. Negative
result logged: _rm_plus_update add.at->bincount was a wash (numpy 2.x
ufunc.at already fast on sorted indices) — reverted. K1 46/46, fast BR
suite 0.142s (p50 0.81ms).

+++ hoyt/profiles.py
@@ -554,8 +554,8 @@ class _NetSigmaProvider:
             anc = bn
             for j in range(eng.cw, 0, -1):
                 wj = waves[j]
-                tj = wj["tile"][anc]
-                sj = wj["pseat"][anc]
+                tj = wj["tile"][anc].astype(np.int64)
+                sj = wj["pseat"][anc].astype(np.int64)
                 k = eng.p0 + j - 1
                 base = 9 * tj
                 blocks[arU, base + (sj - povs) % 4] = 1.0
